requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
stripe==7.8.0
Flask-Mail==0.9.1
requests==2.31.0
authlib==1.3.0

# Dependencias opcionales para Analytics y Reportes
# Para exportación a Excel:
pandas>=2.0.0
openpyxl>=3.1.0

# Para exportación a PDF:
reportlab>=4.0.0

# Módulo certificados (PDF + QR):
weasyprint>=60.0
qrcode[pil]>=7.4
